DTD-2485: add channelIdentifier field to UpdatePlanRequest
