+++ test/uk/gov/hmrc/timetopayproxy/connectors/TtpConnectorSpec.scala
@@ -349,6 +349,7 @@ class TtpConnectorSpec extends PlaySpec with DefaultAwaitTimeout with FutureAwai
             CustomerReference("CustRef1234"),
             PlanId("PlanId1234"),
             UpdateType("Cancel"),
+            None,
             Some(PlanStatus.ResolvedCancelled),
             None,
             None,
@@ -373,6 +374,7 @@ class TtpConnectorSpec extends PlaySpec with DefaultAwaitTimeout with FutureAwai
             CustomerReference("CustRef1234"),
             PlanId("PlanId1234"),
             UpdateType("Cancel"),
+            None,
             Some(PlanStatus.ResolvedCancelled),
             None,
             None,
